Put common parts of texture_zero.spec.ts into a separate file
Saw this MAINTENANCE_TODO so did it
